requirements.txt
Pyrogram==2.0.106
TgCrypto==1.2.5
SQLAlchemy[asyncio]==2.0.37
Alembic==1.9.0
redis[hiredis]==5.2.1
python-dotenv==1.0.1
#aiologger==0.7.0 # need right to set up him for project. not now
aiofiles==24.1.0
asyncpg==0.30.0
black==25.1.0
isort==6.0.0
mypy==1.14.1
autoflake==2.3.1